+++ test/test_variables.py
@@ -3,6 +3,10 @@
 This module aims at testing the correct behavior of the Variables class.
 """
 
+import warnings
+from datetime import datetime
+from zoneinfo import ZoneInfo
+
 import numpy as np
 import pandas as pd
 import pytest
@@ -12,8 +16,11 @@ import xarray.core.utils
 
 import linopy
 from linopy import Model
+from linopy.common import CoordAlignWarning, TimezoneAlignError
 from linopy.testing import assert_varequal
 from linopy.variables import ScalarVariable
+
+UTC = ZoneInfo("UTC")
 
 
 @pytest.fixture
@@ -122,3 +129,48 @@ def test_scalar_variable(m: Model) -> None:
     x = ScalarVariable(label=0, model=m)
     assert isinstance(x, ScalarVariable)
     assert x.__rmul__(x) is NotImplemented  # type: ignore
+
+
+def test_timezone_alignment_with_multiplication() -> None:
+    utc_index = pd.date_range(
+        start=datetime(2025, 1, 1),
+        freq="15min",
+        periods=4,
+        tz=UTC,
+        name="time",
+    )
+    model = Model()
+    series1 = pd.Series(index=utc_index, data=1.0)
+    var1 = model.add_variables(coords=[utc_index], name="var1")
+
+    with warnings.catch_warnings():
+        warnings.simplefilter("error", CoordAlignWarning)
+        expr = var1 * series1
+
+    index: pd.DatetimeIndex = expr.coords["time"].to_index()
+    assert index.equals(utc_index)
+    assert index.tzinfo is UTC
+
+
+def test_timezone_alignment_failure() -> None:
+    utc_index = pd.date_range(
+        start=datetime(2025, 1, 1),
+        freq="15min",
+        periods=4,
+        tz=UTC,
+        name="time",
+    )
+    tz_naive_index = pd.date_range(
+        start=datetime(2025, 1, 1),
+        freq="15min",
+        periods=4,
+        tz=None,
+        name="time",
+    )
+    model = Model()
+    series1 = pd.Series(index=tz_naive_index, data=1.0)
+    var1 = model.add_variables(coords=[utc_index], name="var1")
+
+    with pytest.raises(TimezoneAlignError):
+        # We expect to get a useful error (TimezoneAlignError) instead of a not implemented error falsely claiming that we cannot multiply these types together
+        _ = var1 * series1